langgraph_crawler/requirements.txt
# 核心依赖
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
